Refactor validation schemas and seprate it from types. update API request handling
